refactor: BreadCategory 도메인 분리 (#39)
