+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/breadcategory/repository/BreadCategoryRepository.java
@@ -4,7 +4,7 @@ import static com.breaditnow.domain.global.exception.DomainErrorCode.*;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.breaditnow.domain.domain.product.entity.BreadCategory;
+import com.breaditnow.domain.domain.breadcategory.entity.BreadCategory;
 import com.breaditnow.domain.global.exception.DomainException;
 
 public interface BreadCategoryRepository extends JpaRepository<BreadCategory, Long> {
